I'm trying to fix the gifs
GIFs are really hard to figure out where the end of the file is. If you just look for the EOF marker they can get cut off in the middle so you have to iterate through the file to find the EOF marker where it's actually supposed to be. This is painful.